hf_space/requirements.txt
gradio
tensorflow==2.12.0
Pillow
numpy
huggingface_hub
